core: tools: blueos_startup_update: Check for file before deleting

+++ core/tools/blueos_startup_update/blueos_startup_update.py
@@ -524,6 +524,7 @@ def main() -> int:
 if __name__ == "__main__":
     try:
         main()
-        os.remove(BOOT_LOOP_DETECTOR)
+        if os.path.exists(BOOT_LOOP_DETECTOR):
+            os.remove(BOOT_LOOP_DETECTOR)
     except Exception as error:
         logger.error(f"An error occurred while applying patches: {error}")
